finished the question portion of the application and added sqlite back
When answering a question, the response is recorded with the question on the question-bank page shown with a green or red color depending on the answer provided. Make sure to sign into (username: john password: capstone2022) for testing and answering the questions in the question bank. Also, superuser is (username: user password: capstone2022).